- state tracking

+++ flow.js
@@ -1,94 +1,187 @@
+'use strict';
+
 const flowDebug = require('debug')('flux:flow');
 const cluster = require('cluster');
 const fs = require('fs');
 const path = require('path');
 
 
-var Flow = module.exports = function Flow(flux, obj) {
+/**
+ *	Constructor for Flow
+ *	@constructor
+ *	@param	{Object=}	FLUX
+ */
+var Flow = module.exports = function Flow(FLUX) {
 
 	this.nodes = [];
 	this.connectors = [];
 
-	if (flux) {
+	if (FLUX) {
 
-		this.setFlux(flux);
-
-		if (!this._id) {
-			this._id = flux.generateObjectId();
-		}
-
-		if (obj) {
-
-			if (obj._id) {
-				this._id = obj._id;
-			} else {
-				obj._id = this._id;
-			}
-
-			this.initJson(obj);
-			this.save();
-
-		}
-
-		flux.addFlow(this);
+		this.setFlux(FLUX);
+		this._id = FLUX.generateObjectId();
 
 	}
 
 };
 
 
+/**
+ *	Flow module initialization function
+ *	@static
+ *	@param	{Object}	Flux
+ */
 Flow.init = function(Flux) {
 
 	this.Flux = Flux;
 	Flux.Flow = Flux.prototype.Flow = this;
+	Flux.FlowState = Flux.FlowState = FlowState;
 
 };
 
 
-Flow.initFromPath = function(flowPath, flux) {
+/**
+ *	Init flows from a given path.
+ *	This will parse all json files except for status.json into flows.
+ *	Note that a path cannot be initiated twice because it is used for saveStatuses()
+ *	@static
+ *	@param	{String}	path
+ *	@param	{Object}	flux
+ *	@return {Object.<String, Flow>}	list of flows by their _id
+ */
+Flow.initFromPath = function(flowPath, FLUX) {
 
 	flowDebug(`init flows from ${flowPath}`);
 
-	let flows = [];
+	if (this.flowPath) {
 
+		flowDebug(`cannot init two flow paths. ${this.flowPath} already init`);
+		return;
+
+	}
+
+	Flow.flowPath = flowPath;
+
+	//will hold the flows by their _id
+	let flows = {};
+
+	//get the files in the flowPath
 	fs.readdirSync(flowPath).forEach((file) => {
 
-		var filepath = flowPath + '/' + file;
-		if (fs.statSync(filepath).isFile() && path.extname(filepath) === '.json') {
+		let filepath = flowPath + '/' + file;
 
-			let flow = new Flow(flux);
-			flow.initJson(JSON.parse(fs.readFileSync(filepath)));
-			flows.push(flow);
-			flow.start();
+		//only fetch json files but ignore status.json
+		if (file !== 'status.json' && fs.statSync(filepath).isFile() && path.extname(filepath) === '.json') {
+
+			try {
+
+				let json = JSON.parse(fs.readFileSync(filepath));
+				if (json._id) {
+
+					let flow = new Flow(FLUX);
+					flow.initJson(json);
+					flows[flow._id] = flow;
+
+				}
+
+			} catch (e) {
+				flowDebug(`could not init '${filepath}': ${e}`);
+			}
 
 		}
 
 	});
+
+	//start all flows which had status running before
+	let statuses = this.getStatuses();
+	for (let flowId in statuses) {
+		if (statuses[flowId] && flows[flowId]) {
+			flows[flowId].start();
+		}
+	}
 
 	return flows;
 
 };
 
 
-//set the flux belonging to this flow
-Flow.prototype.setFlux = function(flux) {
+/**
+ *	get all flow statuses
+ *	@static
+ *	@return {Object.<String, Boolean>}	statuses
+ */
+Flow.getStatuses = function() {
 
-	if (flux && flux instanceof Flow.Flux) {
-		return (this.flux = flux);
-	} else {
-		throw new Error('flux must be instanceof Flux');
+	if (!this.flowPath) {
+		return;
+	}
+
+	let statuses = {};
+
+	try {
+		statuses = JSON.parse(fs.readFileSync(`${this.flowPath}/status.json`));
+	} catch (err) {
+		flowDebug(`${this.flowPath}/status.json cannot be opened: ${err}`);
+	}
+
+	return statuses;
+
+};
+
+
+/**
+ *	save statuses
+ *	@static
+ *	@param	{Object.<String, Boolean>}	statuses
+ */
+Flow.saveStatuses = function(statuses) {
+
+	if (!this.flowPath) {
+		return;
+	}
+
+	try {
+		fs.writeFileSync(`${this.flowPath}/status.json`, JSON.stringify(statuses));
+	} catch (err) {
+		flowDebug(`error saving status to file: ${err}`);
 	}
 
 };
 
 
-//init a flow, including all its nodes and connectors, from a json obj
-Flow.prototype.initJson = function(json) {
+/**
+ *	set the FLUX instance belonging to this flow
+ *	@throw	throws if FLUX is not instanceof Flow.Flux
+ *	@param	{Object} FLUX
+ */
+Flow.prototype.setFlux = function(FLUX) {
+
+	if (FLUX && FLUX instanceof Flow.Flux) {
+		return (this.flux = FLUX);
+	} else {
+		throw new Error('FLUX must be instanceof Flux');
+	}
+
+};
+
+
+/**
+ *	init a flow, including all its nodes and connectors, from a json obj
+ *	@param	{Object}	json
+ *	@param	{Boolean}	newFlow	indicates if this is a new flow to be created
+ */
+Flow.prototype.initJson = function(json, newFlow) {
 
 	flowDebug(`initJson on ${json._id}`);
 
 	if (!this.flux) {
-		throw new Error('flux must be set');
+		throw new Error('FLUX must be set');
+	}
+
+	if (!newFlow && json._id) {
+		this._id = json._id;
+	} else {
+		json._id = this._id;
 	}
 
 	this.json = json;
@@ -98,9 +191,15 @@ Flow.prototype.initJson = function(json) {
 	//get the nodes
 	json.nodes.forEach(node => {
 
-		let fluxNode = new this.flux.Node(this.flux.getNodeByName(node.name));
+		let nodeConstr = this.flux.getNodeByName(node.name);
+
+		if (!nodeConstr) {
+			throw (`Node '${node.name}' does not exist`);
+		}
+
+		let fluxNode = new this.flux.Node(nodeConstr);
 		if (!fluxNode) {
-			return;
+			throw (`Could not construct node '${node.name}'`);
 		}
 
 		fluxNode._id = node._id;
@@ -111,11 +210,23 @@ Flow.prototype.initJson = function(json) {
 		this.addNode(fluxNode);
 
 		for (let name in node.inputs) {
+
+			if (!fluxNode.inputs[name]) {
+				throw (`Node '${node.name}' does not have input '${name}'`);
+			}
+
 			fluxNode.inputs[name]._id = node.inputs[name]._id;
+
 		}
 
 		for (let name in node.outputs) {
+
+			if (!fluxNode.outputs[name]) {
+				throw (`Node '${node.name}' does not have output '${name}'`);
+			}
+
 			fluxNode.outputs[name]._id = node.outputs[name]._id;
+
 		}
 
 	});
@@ -124,7 +235,14 @@ Flow.prototype.initJson = function(json) {
 	json.connectors.forEach(connector => {
 
 		var origin = this.getOutputById(connector.origin);
+		if (!origin) {
+			throw (`Cannot find output by id '${connector.origin}'`);
+		}
+
 		var destination = this.getInputById(connector.destination);
+		if (!destination) {
+			throw (`Cannot find input by id '${connector.destination}'`);
+		}
 
 		var fluxConnector = {
 
@@ -138,29 +256,40 @@ Flow.prototype.initJson = function(json) {
 
 	});
 
+	this.flux.addFlow(this);
+
 };
 
 
-//saves a flow to the configured flows directory
-//only works if this is a the master thread
+/**
+ *	saves a flow to the configured flows directory
+ *	only works if this is a the master thread
+ *	@param {Function}	callback
+ */
 Flow.prototype.save = function(callback) {
 
-	if (cluster.isMaster && this._id) {
+	if (cluster.isMaster && this._id && Flow.flowPath) {
 
 		flowDebug(`saving ${this._id}`);
-		fs.writeFile(`./flows/${this._id}.json`, JSON.stringify(this.json), callback);
+		fs.writeFile(`${Flow.flowPath}/${this._id}.json`, JSON.stringify(this.json), callback);
 
 	}
 
 };
 
 
-//not sure if needed?
+/*
 Flow.prototype.addConnector = function(conn) {
-	this.connectors.push(connector);
+	this.connectors.push(conn);
 };
+*/
 
 
+/**
+ *	adds a node to a flow
+ *	@param	{Node}	node
+ *	@return {Node}	node
+ */
 Flow.prototype.addNode = function(node) {
 
 	node.flow = this;
@@ -197,10 +326,20 @@ Flow.prototype.addNode = function(node) {
 	//track output triggers
 	node.prependListener('triggerout', function(output) {
 
-		var d = new Date();
-		var diff = d.getTime() - this._trackerTriggerTime;
+		let d = new Date();
+		let msg;
+
+		if (this._trackerTriggerTime) {
+
+			let diff = d.getTime() - this._trackerTriggerTime;
+			msg = `triggered '${output.name}' in ${diff}ms`;
+
+		} else {
+			msg = `triggered '${output.name}' @ ${d.getHours()}:${d.getMinutes()}:${d.getSeconds()}:${d.getMilliseconds()}`;
+		}
+
 		this.setTracker({
-			message: `triggered '${output.name}' @ ${d.getHours()}:${d.getMinutes()}:${d.getSeconds()}:${d.getMilliseconds()} in ${diff}ms`,
+			message: msg,
 			timeout: 3500
 		});
 
@@ -213,11 +352,21 @@ Flow.prototype.addNode = function(node) {
 };
 
 
+/**
+ *	returns a node from a specific flow by the node._id
+ *	@param {Number}	id	the _id of the node the be found
+ *	@return {Node}	the found node
+ */
 Flow.prototype.getNodeById = function(id) {
 	return this.nodes.find(node => node._id === id);
 };
 
 
+/**
+ *	returns an input for any node by the input._id
+ *	@param {Number}	id	the _id of the nodeInput to be found
+ *	@return {NodeInput}	the found nodeInput
+ */
 Flow.prototype.getInputById = function(id) {
 
 	for (let i = 0; i < this.nodes.length; i++) {
@@ -236,6 +385,11 @@ Flow.prototype.getInputById = function(id) {
 };
 
 
+/**
+ *	returns an output for any node by the output._id
+ *	@param {Number}	id	the _id of the nodeOutput to be found
+ *	@return {NodeOutput}	the found nodeOutput
+ */
 Flow.prototype.getOutputById = function(id) {
 
 	for (let i = 0; i < this.nodes.length; i++) {
@@ -254,63 +408,95 @@ Flow.prototype.getOutputById = function(id) {
 };
 
 
+/**
+ *	saves the status (running or not) for this flow by calling Flow.saveStatuses()
+ *	@param	{Boolean}	running	status of the flow
+ */
+Flow.prototype.saveStatus = function(running) {
+
+	let statuses = Flow.getStatuses();
+	statuses[this._id] = !!running;
+	Flow.saveStatuses(statuses);
+
+};
+
+
+/**
+ *	Starts a flow. Stops it first if it is already running.
+ *	Note that the behaviour is different when called from a worker process
+ *	@return {Promise}
+ */
 Flow.prototype.start = function() {
 
 	if (cluster.isMaster) {
 
-		flowDebug('starting flow from master');
+		return new Promise((resolve, reject) => {
 
-		this.worker = cluster.fork();
-		this.worker.on('message', message => {
+			this.stop().then(() => {
 
-			switch (message.method) {
+				flowDebug('starting flow from master');
 
-				case 'init':
+				//save the status
+				this.saveStatus(true);
 
-					if (this.worker && this.worker.isConnected()) {
+				this.worker = cluster.fork();
+				this.worker.on('message', message => {
 
-						flowDebug('flow/worker has started');
+					switch (message.method) {
 
-						this.worker.send({
-							"method": "start",
-							"flow": this.json
-						});
+						case 'init':
 
-					} else {
-						flowDebug('flow/worker has started, but no such worker in master');
+							if (this.worker && this.worker.isConnected()) {
+
+								flowDebug('flow/worker has started');
+								resolve();
+
+								this.worker.send({
+									"method": "start",
+									"flow": this.json
+								});
+
+							} else {
+
+								flowDebug('flow/worker has started, but no such worker in master');
+								reject(`no such worker in master`);
+
+							}
+
+							break;
+
+						case 'stop':
+
+							this.stop();
+							break;
+
+						case 'broadcastWebSocket':
+
+							if (this.flux && this.flux.webSocketServer) {
+
+								this.flux.webSocketServer.clients.forEach(client => {
+									client.send(JSON.stringify(message.message));
+								});
+
+							}
+
+							break;
+
 					}
 
-					break;
+				});
 
-				case 'stop':
+				this.worker.on('exit', () => {
+					flowDebug(`worker exited`);
+				});
 
-					this.stop();
-					break;
+				this.worker.on('disconnect', () => {
+					flowDebug(`worker disconnected`);
+				});
 
-				case 'broadcastWebSocket':
-
-					if (this.flux && this.flux.webSocketServer) {
-
-						this.flux.webSocketServer.clients.forEach(client => {
-							client.send(JSON.stringify(message.message));
-						});
-
-					}
-
-					break;
-
-			}
+			});
 
 		});
-
-		this.worker.on('exit', () => {
-			flowDebug(`worker exited`);
-		});
-
-		this.worker.on('disconnect', () => {
-			flowDebug(`worker disconnected`);
-		});
-
 
 	} else {
 
@@ -318,13 +504,13 @@ Flow.prototype.start = function() {
 
 		//init any node that wants to
 		this.nodes.forEach(node => {
-			node.emit('init');
+			node.emit('init', new FlowState());
 		});
 
 		//trigger all event objects that are listening
 		this.nodes.forEach(node => {
 			if (node.type === 'event') {
-				node.emit('trigger');
+				node.emit('trigger', new FlowState());
 			}
 		});
 
@@ -333,49 +519,70 @@ Flow.prototype.start = function() {
 };
 
 
+/**
+ *	Stops a flow. Will forcibly kill the flow if it is still running after 5 seconds.
+ *	Note that the behaviour is different when called from a worker process.
+ *	@return {Promise}
+ */
 Flow.prototype.stop = function() {
 
 	if (cluster.isMaster) {
 
-		if (this.worker && this.worker.isConnected()) {
+		return new Promise((resolve, reject) => {
 
-			flowDebug('stopping flow from master');
-			var killTimeout;
+			this.saveStatus(false);
 
-			this.worker.on('disconnect', () => {
+			if (this.worker && this.worker.isConnected()) {
 
-				clearTimeout(killTimeout);
-				flowDebug('worker disconnected from master');
-				this.worker.kill();
-				this.worker = null;
+				flowDebug('stopping flow from master');
+				var killTimeout;
 
-				//cleanup all open statuses
-				this.flux.webSocketServer.clients.forEach(client => {
-					client.send('{\"method\":\"flux.removeAllStatuses\"}');
+				this.worker.on('disconnect', () => {
+
+					clearTimeout(killTimeout);
+					flowDebug('worker disconnected from master');
+
+					if (this.worker) {
+						this.worker.kill();
+					}
+
+					this.worker = null;
+
+					//cleanup all open statuses
+					this.flux.webSocketServer.clients.forEach(client => {
+						client.send('{\"method\":\"flux.removeAllStatuses\"}');
+					});
+
+					resolve();
+
 				});
 
-			});
-
-			this.worker.send({
-				"method": "stop"
-			});
-			this.worker.disconnect();
-
-			//forcibly kill after 5 seconds
-			killTimeout = setTimeout(() => {
-
-				flowDebug('killing worker from master');
-				this.worker.kill('SIGKILL');
-				this.worker = null;
-
-				//cleanup all open statuses
-				this.flux.webSocketServer.clients.forEach(client => {
-					client.send('{\"method\":\"flux.removeAllStatuses\"}');
+				this.worker.send({
+					"method": "stop"
 				});
+				this.worker.disconnect();
 
-			}, 5000);
+				//forcibly kill after 5 seconds
+				killTimeout = setTimeout(() => {
 
-		}
+					flowDebug('killing worker from master');
+					this.worker.kill('SIGKILL');
+					this.worker = null;
+
+					//cleanup all open statuses
+					this.flux.webSocketServer.clients.forEach(client => {
+						client.send('{\"method\":\"flux.removeAllStatuses\"}');
+					});
+
+					resolve();
+
+				}, 5000);
+
+			} else {
+				resolve();
+			}
+
+		});
 
 	} else {
 
@@ -389,3 +596,46 @@ Flow.prototype.stop = function() {
 	}
 
 };
+
+
+/**
+ *	Contains a flow state.
+ *	@constructor
+ *	@class
+ *	@param	{Object.<String, Object>}	states
+ */
+function FlowState(states = {}) {
+
+	/**
+	 *	Sets a state for a given node.
+	 *	Freezes the object to disallow any future adjustments.
+	 *	@param	{Node}	node
+	 *	@param	{Object}	obj
+	 */
+	this.set = function(node, obj) {
+
+		Object.freeze(obj);
+		states[node._id] = obj;
+
+	};
+
+	/**
+	 *	Gets a state for a given node.
+	 *	@param	{Node}	node
+	 *	@return	{Object}
+	 */
+	this.get = function(node) {
+		return states[node._id];
+	};
+
+	/**
+	 *	Splits the flowState into a new flowState
+	 *	@return	{FlowState}	the new flowState
+	 */
+	this.split = function() {
+		return new FlowState(Object.assign({}, states));
+	};
+
+	Object.freeze(this);
+
+}
